Working out Tier Two Module concept
The main idea around tier two modules is to make something that is kindof like regular modules, but they are allowed a little bit of dependency, as a treat <3

Also started working out the first tier two module which is an extension of the Lock & Key module. Only has one dependency so far, but I expect to have more later in this one. Hopefully the tier two modules help in the way I expect and don't introduce any new problems