Added intelligent static mode settings
Allows for pickup items to opt into / out of  setting static mode once the physics body starts sleeping. It also can start static if desired. The default mode is the object is a rigid body until the player has picked it up, from then on if it is being simulated, the body becomes static once it starts sleeping. Later this property will be persistent which lets the player place furniture easily