Major Tweaks!
Started adding some dialogic content to make sure that integrates well but it actually doesn't! I need a system to take control away from the player while a dialogue is playing. This led me to start refactoring the player controller code into a hierarchical state machine, inspired by the work of "The Shaggy Dev" on youtube.