Made settings save to disk and reload when next opened
Made some changes to the SaveData singleton to allow for generic save/load/delete operations as long as a path is provided. Tweaked the Subtitles because it crashes if the Settings are changed at startup (before children added). Also ensured the settings UI elements match loaded settings and reload the settings when changed. Because of this system, even reverting to fullscreen on startup is supported